Moved to Rx\Exception namespace and added base Exception
