+++ src/Exception/Exception.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rx\Exception;
+
+class Exception extends \Exception
+{
+}
