+++ src/Exception/TimeoutException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rx\Exception;
+
+class TimeoutException extends Exception
+{
+}
