--- src/TimeoutException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Rx;
-
-class TimeoutException extends \Exception
-{
-}
